Minor updates
fixed an issue when deleting a post.
fixed css hover on three 'no posts' classes so that nothing happens when you hover over them
fixed the home page to only display the sign in/register when there is no user logged in.
removed the about file as it is never used.